Place 모든 기능 수정 및 더미데이터 추가
