+++ src/main/java/com/pickyfy/pickyfy/service/AdminService.java
@@ -6,9 +6,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface AdminService {
-    Long createPlace(PlaceCreateRequest request, List<MultipartFile> imageList);
-    Long updatePlace(Long placeId, PlaceCreateRequest request, List<MultipartFile> imageList);
-    void deletePlace(Long placeId);
-    void deletePlaceImages(Long placeImageId);
+
     void logout(String accessToken);
 }
