+++ src/main/java/com/pickyfy/pickyfy/service/PlaceServiceImpl.java
@@ -2,13 +2,17 @@ package com.pickyfy.pickyfy.service;
 
 import com.pickyfy.pickyfy.apiPayload.code.status.ErrorStatus;
 import com.pickyfy.pickyfy.domain.*;
+import com.pickyfy.pickyfy.web.dto.request.PlaceCreateRequest;
 import com.pickyfy.pickyfy.web.dto.response.PlaceSearchResponse;
 import com.pickyfy.pickyfy.repository.*;
+import jakarta.persistence.EntityExistsException;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -27,6 +31,9 @@ public class PlaceServiceImpl implements PlaceService {
     final private CategoryRepository categoryRepository;
     final private MagazineRepository magazineRepository;
     final private PlaceMagazineRepository placeMagazineRepository;
+    final private S3Service s3Service;
+    final private PlaceCategoryRepository placeCategoryRepository;
+
 
     /**
      * 특정 유저가 저장한 Place 전체 조회
@@ -122,6 +129,12 @@ public class PlaceServiceImpl implements PlaceService {
     }
 
 
+    /**
+     * 유저 Place 저장 및 저장취소 (toggle)
+     * @param userId
+     * @param placeId
+     * @return
+     */
     @Transactional
     public boolean togglePlaceUser(Long userId, Long placeId) {
 
@@ -159,4 +172,160 @@ public class PlaceServiceImpl implements PlaceService {
             return true;
         }
     }
+
+    /**
+     * 관리자 기능 (Place 생성)
+     * @param request
+     * @param imageList
+     * @return
+     */
+    @Transactional
+    public Long createPlace(PlaceCreateRequest request, List<MultipartFile> imageList) {
+
+        if (placeRepository.existsPlaceByName(request.name())) {
+            throw new EntityExistsException("Place Already exists");
+        }
+
+        Category category = categoryRepository.findById(request.categoryId()).orElseThrow(() ->
+                new EntityNotFoundException("Category not found"));
+
+        Magazine magazine = magazineRepository.findById(request.magazineId()).orElseThrow(() ->
+                new EntityNotFoundException("Magazine not found"));
+
+        // 1. Place 먼저 저장
+        Place newPlace = Place.builder()
+                .name(request.name())
+                .longitude(request.longitude())
+                .latitude(request.latitude())
+                .address(request.address())
+                .instagramLink(request.instagramLink())
+                .naverplaceLink(request.naverPlaceLink())
+                .shortDescription(request.shortDescription())
+                .build();
+
+        newPlace = placeRepository.save(newPlace); // 먼저 저장
+
+        // 2. 저장된 newPlace를 이용하여 PlaceCategory, PlaceMagazine 저장
+        PlaceCategory newPlaceCategory = PlaceCategory.builder()
+                .category(category)
+                .place(newPlace)
+                .build();
+
+        placeCategoryRepository.save(newPlaceCategory);
+
+        PlaceMagazine newPlaceMagazine = PlaceMagazine.builder()
+                .magazine(magazine)
+                .place(newPlace)
+                .build();
+
+        placeMagazineRepository.save(newPlaceMagazine);
+
+        // 3. 이미지 저장 로직
+        List<PlaceImage> placeImages = new ArrayList<>();
+        int maxImages = Math.min(imageList.size(), 5);
+
+        for (int i = 0; i < maxImages; i++) {
+            String imageUrl = s3Service.upload(imageList.get(i));
+            PlaceImage newPlaceImage = PlaceImage.builder()
+                    .place(newPlace)
+                    .url(imageUrl)
+                    .sequence(i)
+                    .build();
+            placeImages.add(newPlaceImage);
+        }
+
+        newPlace.getPlaceImages().addAll(placeImages);
+        placeRepository.save(newPlace); // 최종 저장
+
+        return newPlace.getId();
+    }
+
+
+
+    /**
+     * 관리자 기능(Place 수정)
+     * @param placeId
+     * @param request
+     * @param imageList
+     * @return
+     */
+    @Override
+    @Transactional
+    public Long updatePlace(Long placeId, PlaceCreateRequest request, List<MultipartFile> imageList) {
+        Place place = placeRepository.findById(placeId)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorStatus.PLACE_NOT_FOUND.getMessage()));
+
+        place.updatePlace(request.name(), request.address(), request.shortDescription(),
+                request.instagramLink(), request.naverPlaceLink(), request.latitude(), request.longitude() );
+
+        Category category = categoryRepository.findById(request.categoryId()).get();
+
+        Magazine magazine = magazineRepository.findById(request.magazineId()).get();
+
+        PlaceCategory placeCategory = placeCategoryRepository.findById(placeId).get();
+
+        PlaceMagazine placeMagazine = placeMagazineRepository.findById(placeId).get();
+
+        placeCategory.updatePlaceCategory(place, category);
+        placeMagazine.updatePlaceMagazine(place, magazine);
+
+        place.updateImages(imageList, s3Service);
+        return place.getId();
+    }
+
+
+    /**
+     * Place 삭제
+     * @param placeId
+     */
+    @Override
+    @Transactional
+    public void deletePlace(Long placeId) {
+        Place place = placeRepository.findById(placeId)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorStatus.PLACE_NOT_FOUND.getMessage()));
+
+        // 1. PlaceCategory 삭제
+        PlaceCategory placeCategories = placeCategoryRepository.findByPlaceId(placeId);
+        placeCategoryRepository.delete(placeCategories);
+
+        // 2. PlaceMagazine 삭제
+        PlaceMagazine placeMagazines = placeMagazineRepository.findByPlaceId(placeId);
+        placeMagazineRepository.delete(placeMagazines);
+
+        // 3. PlaceSavedPlace 삭제
+        PlaceSavedPlace placeSavedPlaces = placeSavedPlaceRepository.findByPlaceId(placeId);
+        placeSavedPlaceRepository.delete(placeSavedPlaces);
+
+        // 4. SavedPlace 삭제 (placeSavedPlace가 먼저 삭제된 후 가능)
+
+        savedPlaceRepository.delete(placeSavedPlaces.getSavedPlace());
+
+
+        // 5. PlaceImage 삭제 (S3에서도 삭제)
+        for (PlaceImage image : place.getPlaceImages()) {
+            s3Service.removeFile(image.getUrl());
+        }
+        placeImageRepository.deleteAll(place.getPlaceImages());
+
+        // 6. Place 삭제
+        placeRepository.delete(place);
+    }
+
+
+
+    /**
+     * PlaceImage 삭제
+     * @param placeImageId
+     */
+    @Override
+    @Transactional
+    public void deletePlaceImages(Long placeImageId) {
+        PlaceImage placeImage = placeImageRepository.findById(placeImageId)
+                .orElseThrow(() -> new EntityNotFoundException(ErrorStatus.IMAGE_INVALID.getMessage()));
+
+        s3Service.removeFile(placeImage.getUrl());
+        placeImageRepository.delete(placeImage);
+    }
+
+
 }
